fix SessionHandler getLoggedInResidentAccount integrity issue
integrity is ensured by the code using the funciton. Always loading the latest ResidentAccount is unnecessary and inefficient because most of the time what it uses is to get the associated room and floor which will not change for a resident as of now. So there is no need to load it from the DB for this use case. But when other details of the logged in ResidentAccount are to be used, then it is loaded from the DB to ensure integrity.